fixed bug for event products management

+++ EventsDB/dbo/Store Procedures/Products/Products_Insert.sql
@@ -2,7 +2,7 @@
     @VenueId INT,
     @HierarchyId INT,
     @OrganiserId INT,
-    @Price INT
+    @Price DECIMAL(18,2)
 AS
 
 INSERT INTO dbo.Products (VenueId ,HierarchyId ,OrganiserId ,Price)
